"Merge" from ZB570TL_MR7_kernel-3.18.tar.gz
"Version WW MR7"

cd93e3145fd985af9b93e6cb896387d8b5aea503

This was imported on top of running `git am` for the output of
`git format-patch --stdout v3.18.35..v3.18.79` to reduce the diff size.

+++ drivers/misc/mediatek/teei/V1.0/tz_driver/include/teei_common.h
@@ -1,3 +1,17 @@
+/*
+ * Copyright (c) 2015-2017 MICROTRUST Incorporated
+ * All Rights Reserved.
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * version 2 as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ */
+
 
 #ifndef __TEEI_COMMON_H_
 #define __TEEI_COMMON_H_
@@ -11,9 +25,9 @@
  * @brief Command status
  */
 enum teei_cmd_status {
-    TEEI_STATUS_INCOMPLETE = 0,
-    TEEI_STATUS_COMPLETE,
-    TEEI_STATUS_MAX  = 0x7FFFFFFF
+	TEEI_STATUS_INCOMPLETE = 0,
+	TEEI_STATUS_COMPLETE,
+	TEEI_STATUS_MAX  = 0x7FFFFFFF
 };
 
 
@@ -21,8 +35,8 @@ enum teei_cmd_status {
  * @brief Parameters type
  */
 enum teeic_param_type {
-    TEEIC_PARAM_IN = 0,
-    TEEIC_PARAM_OUT
+	TEEIC_PARAM_IN = 0,
+	TEEIC_PARAM_OUT
 };
 
 /**
@@ -38,15 +52,15 @@ struct teeic_notify_data {
 
 
 enum teeic_param_value {
-    TEEIC_PARAM_A = 0,
-    TEEIC_PARAM_B
+	TEEIC_PARAM_A = 0,
+	TEEIC_PARAM_B
 };
 
 enum teeic_param_pos {
-    TEEIC_PARAM_1ST = 0,
-    TEEIC_PARAM_2ND,
-    TEEIC_PARAM_3TD,
-    TEEIC_PARAM_4TH
+	TEEIC_PARAM_1ST = 0,
+	TEEIC_PARAM_2ND,
+	TEEIC_PARAM_3TD,
+	TEEIC_PARAM_4TH
 };
 
 
@@ -55,12 +69,12 @@ enum teeic_param_pos {
  */
 struct teei_encode_meta {
 	int type;
-	int len;			/* data length */
-	unsigned long usr_addr;		/* data address in user space */
-	int ret_len;			/* return sizeof data */
-	int value_flag;			/* value of a or b */
-	int param_pos;			/* param order */
-	int param_pos_type;		/* param type */
+	int len;                        /* data length */
+	unsigned  long long usr_addr;		/* data address in user space */
+	int ret_len;                    /* return sizeof data */
+	int value_flag;                 /* value of a or b */
+	int param_pos;                  /* param order */
+	int param_pos_type;             /* param type */
 };
 
 /**
